refactor: 타입 파일 분리 - 기존에 한 파일에 몰아두었던 타입들을 사용에 맞게 별도의 파일로 분리 - 코드의 가독성과 유지보수성 향상
