+++ src/common/components/Buttons/ProfileEditButton.tsx
@@ -1,4 +1,4 @@
-import { IBaseButtonStyleProps } from '@/common/types/children.interface';
+import { IBaseButtonStyleProps } from "@/common/types/button.interface";
 
 export default function ProfileEditButton({ className, children }: IBaseButtonStyleProps) {
 	return <button className={className}>{children}</button>;
