+++ src/common/components/LiveStatusContainer.tsx
@@ -1,13 +1,7 @@
 import { UserIcon } from '@heroicons/react/24/solid';
-import React from 'react';
+import { ILiveStatusContainerProps } from '../types/container.interface';
 
-interface ILiveStatusContainerProps {
-	watchingCount: number;
-}
-
-export default function LiveStatusContainer({
-	watchingCount,
-}: ILiveStatusContainerProps) {
+export default function LiveStatusContainer({ watchingCount }: ILiveStatusContainerProps) {
 	return (
 		<div className={`absolute left-2 top-2 flex gap-2 lg:text-sm 2xl:text-xl`}>
 			<div className="center bg-mainRed rounded-lg p-1 font-semibold">LIVE</div>
