+++ src/common/types/handler.type.ts
@@ -1,0 +1,1 @@
+export type TInputChangeEvent = React.ChangeEvent<HTMLInputElement>;
