+++ src/common/types/image.interface.ts
@@ -1,0 +1,5 @@
+export interface ICircleImageProps {
+	src: string;
+	size?: 'm' | 'l';
+	isStreaming?: boolean;
+}
