+++ src/common/types/object.interface.ts
@@ -1,0 +1,11 @@
+export interface ITheme {
+	id: string;
+	name: string;
+	icon: string;
+}
+
+export interface IMenuItem {
+	anchor: string;
+	name: string;
+	icon: React.ReactNode;
+}
