+++ src/common/types/search.interface.ts
@@ -1,0 +1,8 @@
+import { TInputChangeEvent } from "./handler.type";
+
+export interface ISearchBarProps {
+	inputText: string;
+	onChangeInput: (event: TInputChangeEvent) => void;
+	onClickReset: () => void;
+  onClickSearch: () => void;
+}
